Added actual screen scrolling relative to player
Other changes such as minor sprite edits, filename changes, cleanup of code with the 'walkaround v2.py' file. walkaround.py will no longer be updated but will be kept around for a short period of time for reference purposes.

Also making progress towards tilemaps and collision detection that is currently not a thing.